fix(engine): BigDecimal overload trap, parser backtracking, Iverson recursion, set ops, formatting
Critical bug fixes in the C++ calc engine (calc-core):

1. BigDecimal(int) overload-resolution trap (CRASH FIX):
   - BigDecimal(0) matched 'explicit BigDecimal(int precision)' instead of
     'BigDecimal(int64_t value, ...)', creating MPFR with precision=0 →
     assertion abort on every arithmetic operation.
   - Fix: removed the 'explicit BigDecimal(int precision)' constructor
     entirely. Added static factory 'BigDecimal::withPrecision(int)' for
     precision construction. All int literals now correctly match the
     int64_t value constructor via implicit conversion.
   - Updated all 11 source files to use withPrecision() for precision-intent
     and int64_t() cast for value-intent calls.

2. Parser function-call vs definition backtracking (CRASH FIX):
   - tryParseParamList() threw exceptions on non-identifier content (e.g.
     'abs(-5)' → '-5' is not a valid param name → throw). This made ALL
     function calls fail with 'Expected parameter name in function definition'.
   - Fix: tryParseParamList now returns {valid=false} instead of throwing.
     parseDefinition backtracks cleanly and falls through to parseLogicalOr.
   - Also fixed: function definition path now consumes ')' before checking
     ':=' (was missing, so 'f(n) := body' also failed).

3. Iverson bracket infinite recursion (SEGFAULT FIX):
   - NodeType::IVERSON case called evaluateIverson(node) on the IVERSON
     node itself → evaluate(node) → re-enters IVERSON case → stack overflow.
   - Fix: pass *node.children[0] (the predicate) to evaluateIverson.

4. BigDecimal toString() negative-number formatting:
   - mpfr_get_str prepends '-' for negatives, but the decimal-point insertion
     code didn't account for the sign character, producing '-.4' instead of
     '-4' and '-.2' instead of '-2'.
   - Fix: offset the insertion position by signLen (1 if negative, 0 if not)
     in all three format modes (FLOATING_POINT, SCIENTIFIC, SIGNIFICANT_DIGITS).

5. Set difference operator '\' (NEW FEATURE per spec):
   - Added BinaryOp::SET_DIFF to the enum.
   - Parser: '\' is handled at the set-ops level (precedence 7) alongside
     cap/cup, matching the spec.
   - Evaluator: evaluateToSet handles SET_DIFF via CalcSet::difference().
   - CalcSet: added SetOp::DIFFERENCE + difference() method.

6. 'in' operator precedence (SPEC COMPLIANCE):
   - Moved 'in' from parseSetOps (level 7) to parseComparison (level 8),
     matching the spec: relations =,<,>,<=,>=,!=,in are all at level 8.
   - Also made comparisons non-chainable (only one per level, per spec).

7. Predefined sets Real/Rational/Quotient/Integer/Zahlen (NEW FEATURE):
   - Added CalcSet::PredefinedSet enum {REAL, RATIONAL, INTEGER}.
   - CalcSet::contains() checks: REAL → finite, RATIONAL → finite,
     INTEGER → isInteger().
   - Evaluator: VARIABLE_REF nodes with these names are recognized in
     evaluateToSet() as predefined sets, not looked up as variables.
   - Set membership/cap/cup/diff operations now moved BEFORE generic
     left/right number evaluation in evaluateBinary(), so set-typed
     right operands don't crash the evaluator.

Verified with 40+ test expressions covering:
  arithmetic, precedence, power (right-assoc), modulo, abs/floor/ceil,
  gcd/lcm, sqrt, trig, sum/prod, factorial, hex/bin literals, variable
  and function definition, relation operators, logic operators, Iverson
  brackets, set difference, set membership, predefined sets, intervals.

+++ calc-core/src/main/cpp/src/crypto_random.cpp
@@ -73,11 +73,11 @@ BigDecimal CryptoRandom::uniform(const BigDecimal& minimum,
 
     // Convert random bytes to a BigDecimal in [0, 1)
     // fraction = sum(randBytes[i] * 256^(bytes-1-i)) / 256^bytes
-    BigDecimal fraction(0);
-    BigDecimal scale(1);
+    BigDecimal fraction(int64_t(0));
+    BigDecimal scale(int64_t(1));
     for (int i = bytes - 1; i >= 0; i--) {
         fraction = fraction.add(BigDecimal(static_cast<int64_t>(randBytes[i])).mul(scale));
-        scale = scale.mul(BigDecimal(256));
+        scale = scale.mul(BigDecimal(int64_t(256)));
     }
     fraction = fraction.div(scale);
